Move creator's PID check to is_bkp_creator_alive() .
Strictly speaking, this is the same. But now i can say, that you can't assume,
that is_bkp_creator_alive() function returns PID: it may be anything else. The
only proposition, which is true: if it returns something, creator still alive.

So, this is just to ensure smoother (possible) transition to new backup file
name format, which does not include creator's PID at all.